Improved scroll and generic UI time series
