+++ services/ui/src/app/time_series/favicon.ico
