+++ services/ui/src/app/time_series/globals.css
