+++ services/ui/src/app/time_series/layout.tsx
