+++ services/ui/src/app/time_series/page.tsx
